Introduce new SensorSuite and Aggregator
To encourage greater use of the Raspberry Pi, break the sensors into
multiple projects/binaries that can be deployed independently. Add a
publisher that can collect data from the temperature and radiation
sensors and push it up to the new aggregator.

Replace the old gateway with a new Aggregator. The old gateway required
knowing ahead of time how to reach the sensors, and also would not work
across NAT (it tried to connect to the sensors), which precluded hosting
the gateway on AWS, but hosting the Pi's at a conference.

The new aggregator dynamically accepts as many sensors as try to connect
to it. It will automatically average the data and relay it to the game
controller.

One other important change is that it is now expected that there is a
single "solar flare" sensor -- a single source of truth for that data
point. The Solar Flare is really an input to the other sensors, so it
does not make sense to have that be randomly generated for each team,
and then to be "averaged" in any way. Either there is a solar flare that
should effect the trends, or there is not. Therefore, the aggregator
queries the solarflare endpoint directly, and it is expected that the
temperature and radiation sensors are querying the same solar flare
endpoint as well.